feat: shpimets_form

+++ Proyecto_web/database/migrations/2024_08_23_210228_create_shipments_table.php
@@ -26,7 +26,7 @@ class CreateShipmentsTable extends Migration
             $table->date('departureDate');
             $table->date('deliveryDate')->nullable();
             $table->string('status');
-            $table->integer('cost');
+            $table->double('cost');
             $table->string('recipientName');
             $table->timestamps();
         });
